+++ backend/src/main/java/site/coduo/pairroomhistory/service/dto/PairRoomHistoryReadResponse.java
@@ -1,4 +1,4 @@
-package site.coduo.pairroomhistory.dto;
+package site.coduo.pairroomhistory.service.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
